Added more tests.
