+++ Tests/VHDLParsingTests/VHDLFileTests.swift
@@ -276,50 +276,6 @@ final class VHDLFileTests: XCTestCase {
         XCTAssertEqual(VHDLFile(rawValue: raw), file)
     }
 
-    /// Test invalid `init(rawValue:)` returns nil.
-    func testInvalidRawValueInit() {
-        XCTAssertNil(VHDLFile(rawValue: ""))
-        XCTAssertNil(VHDLFile(rawValue: "library IEEE"))
-    }
-
-    /// Test invalid entity in raw value init.
-    func testInvalidEntityRawValueInit() {
-        let raw = """
-        entity TestEntity is
-            port(
-                clk: in std_logic
-            );
-        end TestEntity;
-        """
-        XCTAssertEqual(VHDLFile(rawValue: raw), VHDLFile(architectures: [], entities: entities, includes: []))
-        let raw2 = """
-        entity is
-            port(
-                clk: in std_logic
-            );
-        end TestEntity;
-        """
-        XCTAssertNil(VHDLFile(rawValue: raw2))
-        let raw3 = """
-        entity TestEntity
-            port(
-                clk: in std_logic
-            );
-        end TestEntity;
-        """
-        XCTAssertNil(VHDLFile(rawValue: raw3))
-        let raw4 = """
-        entity TestEntity is
-            port(
-                clk: in std_logic
-            );
-        endTestEntity;
-        """
-        XCTAssertNil(VHDLFile(rawValue: raw4))
-        XCTAssertNil(VHDLFile(rawValue: ""))
-        XCTAssertNil(VHDLFile(rawValue: "abc"))
-    }
-
     // swiftlint:disable function_body_length
 
     /// Test architecture raw value init.
@@ -408,6 +364,16 @@ final class VHDLFileTests: XCTestCase {
             VHDLFile(rawValue: raw), VHDLFile(architectures: [], entities: [entity1, entity2], includes: [])
         )
         XCTAssertEqual(VHDLFile(rawValue: raw)?.rawValue, raw)
+        let raw2 = """
+        entity TestEntity is
+            port(
+                clk: in std_logic
+            );
+        end TestEntity;
+        """
+        XCTAssertEqual(
+            VHDLFile(rawValue: raw2), VHDLFile(architectures: [], entities: entities, includes: [])
+        )
     }
 
     /// Test multiple architectures
@@ -495,74 +461,24 @@ final class VHDLFileTests: XCTestCase {
         XCTAssertEqual(VHDLFile(rawValue: raw2)?.rawValue, raw2 + "\n")
     }
 
-    /// Test invalid package init.
-    func testInvalidPackageInit() {
-        let raw3 = """
-        package Package1 iss
-            constant high: std_logic := '1';
-            type Record1_t iss record
-                a: std_logic;
-                b: std_logic;
-            end record Record1_t;
-            type xs iss std_logic_vector(3 downto 0);
-        end package Package1;
-        """
-        XCTAssertNil(VHDLFile(rawValue: raw3))
-        let raw4 = """
-        package 2Package1 is
-            constant high: std_logic := '1';
-            type Record1_t is record
-                a: std_logic;
-                b: std_logic;
-            end record Record1_t;
-            type xs is std_logic_vector(3 downto 0);
-        end package 2Package1;
-        """
-        XCTAssertNil(VHDLFile(rawValue: raw4))
-        let raw5 = """
-        package Package1 iss
-            constant high: std_logic := '1';
-            type Record1_t is record
-                a: std_logic;
-                b: std_logic;
-            end record Record1_t;
-            type xs is std_logic_vector(3 downto 0);
-        end package Package1;
-        """
-        XCTAssertNil(VHDLFile(rawValue: raw5))
-        XCTAssertNil(VHDLFile(rawValue: "package is"))
-    }
-
-    /// Test invalid package body init.
-    func testInvalidPackageBodyInit() {
+    /// Test multiple package bodies.
+    func testMultiplePackageBody() {
         let raw = """
-        package body Package1
-            constant low: std_logic := '0';
-            constant zeros: xs := x"0";
-        end package body Package1;
-        """
-        XCTAssertNil(VHDLFile(rawValue: raw))
-        let raw2 = """
-        package body is
-            constant low: std_logic := '0';
-            constant zeros: xs := x"0";
-        end package body Package1;
-        """
-        XCTAssertNil(VHDLFile(rawValue: raw2))
-        let raw3 = """
         package body Package1 is
             constant low: std_logic := '0';
             constant zeros: xs := x"0";
-        end package body Package2;
-        """
-        XCTAssertNil(VHDLFile(rawValue: raw3))
-        let raw4 = """
+        end package body Package1;
+
         package body Package1 is
             constant low: std_logic := '0';
-            constants zeros: xs := x"0";
+            constant zeros: xs := x"0";
         end package body Package1;
+
         """
-        XCTAssertNil(VHDLFile(rawValue: raw4))
+        XCTAssertEqual(
+            VHDLFile(rawValue: raw),
+            VHDLFile(architectures: [], entities: [], includes: [], packageBodies: bodies + bodies)
+        )
     }
 
 }
